go: Prototype some improvements to fetch performance.
Fetch performance is currently hindered by two major things:

1) We fetch level-by-level, which causes choke-points at every commit in a long
serial chain of commits.

2) We fetch every remote ref one by one, and update remote tracking branches
one by one. Which causes remotes with thousands of branches to perform very
poorly.

In addition, the implementation leaves a lot of perf on the table with more
minor implementation improvements as well, such as CPU-bound value walking.

This branch is an exploration of how much faster fetch can be if we address
some of these issues. Because it is an exploration, the implementation is very
adhoc and it breaks a lot of encapsulation and abstraction boundaries
throughout the value/store/nbs/datas layers.

To see the newer performance, run `dolt jankyfetch [remote]`.